# Conflicts:
#	Jumpscale/builders/db/BuilderRedis.py

+++ Jumpscale/servers/etcd/EtcdServer.py
@@ -12,6 +12,8 @@ class EtcdServer(j.application.JSBaseClass):
         :return: tmux pane
         :rtype: tmux.Pane
         """
+        if not j.core.tools.cmd_installed("etcd"):
+            raise RuntimeError("install etcd: 'j.builders.db.etcd.install()'")
         cmd = j.sal.fs.joinPaths(j.core.dirs.BINDIR, "etcd")
         if config_file:
             cmd += " --config-file %s" % config_file
@@ -25,6 +27,8 @@ class EtcdServer(j.application.JSBaseClass):
         :param sig: signal, if not given, SIGKILL will be used
         :type sig: int, defaults to None
         """
+        if not j.core.tools.cmd_installed("etcd"):
+            raise RuntimeError("install etcd: 'j.builders.db.etcd.install()'")
         if pid:
             j.sal.process.kill(pid, sig)
         else:
